Mark Row and Cell classes as open so that they can be subclassed
The classes in question were marked as `public`, which is not
a problematic when the package is imported directly into a project.

When the package is installed using a package manager (SPM or other)
these classes couldn't be subclassed because of the `public` access
level. Marking them as `open` solves this problem.